python/requirements.txt
# Realistic Shadow Generator - Python Dependencies

# Core image processing
numpy>=1.20.0
Pillow>=9.0.0
scipy>=1.7.0

# GUI (optional - for ui.py)
PyQt6>=6.4.0

# Background removal (optional - for auto-cutout)
# rembg>=2.0.0  # Uncomment if you want ML-based background removal
